+++ config/settings_ci.py
@@ -73,3 +73,5 @@ LOGIN_REDIRECT_URL = "/"  # this is the name of the url
 
 STATIC_ROOT = "/var/www/osidb/static/"
 STATIC_URL = "/static/"
+
+INSTALLED_APPS += ["osidb.tests"]